Nivel 6
Se crean eventos para abrir puertas
Se actualiza iconos en interfaz
Se actualiza texto en interfaz
Se modifica la generación dinámica de iconos con eventos
Se crea escena Portada
Se carga siguiente escena
Se prueba del prototipo

+++ Assets/Scripts/Jugador/PerfilJugador.cs
@@ -52,8 +52,8 @@ public class PerfilJugador : ScriptableObject
     //vida
     [SerializeField]
     [Range(5, 10)]
-    private float vida = 5f;
-    public float Vida { get => vida; set => vida = value; }
+    private int vida = 5;
+    public int Vida { get => vida; set => vida = value; }
 
     // variables SFX serializadas y encapsuladas
     [Header("Configuraciones SFX")]
